requirements.txt
# CORE DASHBOARD
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0

# DATABASE
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# UTILITIES
python-dotenv==1.0.0
tiktoken==0.7.0
pytz==2023.3

# LANGCHAIN 0.2 (Pydantic v1)
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langsmith==0.1.100
langchain-openai==0.1.22
langchain-groq==0.1.9

# GROQ CLIENT
groq==0.4.2
httpx==0.25.2
pydantic==1.10.14
